pyproject.toml
[project]
name = "pdf-context"
version = "0.1.0"
description = "PDF Context Retrieval"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.4",
    "faiss-cpu>=1.12.0", 
    "langchain-core>=1.0.3", 
    "streamlit>=1.51.0", 
    "langchain-ollama>=1.0.0"
]

[project.optional-dependencies]
test = ["pytest>=9.0.1",]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "UP", # pyupgrade
]

ignore = [
    "E501", # line-too-long
]